backend/requirements.txt
fastapi
uvicorn
langgraph
yfinance
pydantic
google-generativeai
python-dotenv
